added drivecharacterization
moved ff characterization and wheel radius characterization from drive commands --> drive characterization
